+++ fritzconnection/lib/fritzwlan.py
@@ -26,9 +26,37 @@ else:
 # important: don't set an extension number here:
 SERVICE = 'WLANConfiguration'
 DEFAULT_PASSWORD_LENGTH = 12
+WPA_SECURITY = 'WPA'
+NO_PASS = 'nopass'
 
 
-def get_wifi_qr_code(instance, kind='svg', security=None, hidden=False):
+def get_beacon_security(instance, security):
+    """
+    Returns the beacon-security as a string based on the security
+    argument. Possible return values are 'nopass' and 'WPA'. If the
+    security is None or an empty string, the function tries to find the
+    proper security setting ('nopass'|'WPA'). If security is neither
+    None nor an empty string, the value is returned as is.
+
+    This function is not intended to get called directly.
+
+    .. versionadded:: development
+    """
+    if not security:
+        info = instance.get_info()
+        beacontypes = set(info["NewX_AVM-DE_PossibleBeaconTypes"].split(","))
+        beacontypes -= set(('None', 'OWETrans'))
+        beacontype = info["NewBeaconType"]
+        if beacontype in beacontypes:
+            security = WPA_SECURITY
+        else:
+            security = NO_PASS
+    return security
+
+
+def get_wifi_qr_code(instance, kind='svg',
+                     security=None, hidden=False,
+                     scale=4):
     """
     Returns a file-like object providing a bytestring representing a
     qr-code for wlan access. `instance` is a FritzWLAN or FritzGuestWLAN
@@ -38,22 +66,22 @@ def get_wifi_qr_code(instance, kind='svg', security=None, hidden=False):
     This function is not intended to get called directly. Instead it is
     available as a method on FritzWLAN instances (as well as on
     subclasses like FritzGuestWLAN) if the third party package `segno`
-    is installed (pip install segno).
+    is installed.
 
     Consider `guest_wlan` is a FritzGuestWLAN instance, then the
     following code will return a file like object with the qr-code image
-    data in png-format:
+    data in png-format: ::
 
-    >>> stream = guest_wlan.get_wifi_qr_code(kind='png')
+        stream = guest_wlan.get_wifi_qr_code(kind='png')
 
     The stream can get used anywhere, where a file like object is
     expected, i.e. writing the content to a file (Note: the suffix must
-    match the kind of the qr-code format):
+    match the kind of the qr-code format): ::
 
-    >>> with open('qr_code.png', 'wb') as fobj:
-    >>> ....fobj.write(stream.read())
+        with open('qr_code.png', 'wb') as fobj:
+            fobj.write(stream.read())
 
-    If the `segno` is not installed the call will trigger an
+    If `segno` is not installed the call will trigger an
     AttributeError when called on an instance and a NameError when
     called directly.
 
@@ -61,20 +89,29 @@ def get_wifi_qr_code(instance, kind='svg', security=None, hidden=False):
 
     The parameters `security` and `hidden` allow to forward these
     informations to the `segno` library. `security` is `None` or a
-    string like `WPA2`. `hidden` is a boolean value indicating the
+    string like `WPA`. `hidden` is a boolean value indicating the
     visibility of the network .
 
     .. versionadded:: 1.9.1
 
+    If the `security` is `None` (default) the beacontype of the network
+    is used to set the WLAN-Type accordingly. If the value of `security`
+    is something else, this value gets used.
+
+    `scale` defines the size of the produced qr-code. Default value is 4.
+
+    .. versionadded:: development
+
     """
     stream = io.BytesIO()
+    security = get_beacon_security(instance, security)
     qr_code = segno.helpers.make_wifi(
         ssid=instance.ssid,
         password=instance.get_password(),
         security=security,
         hidden=hidden
     )
-    qr_code.save(out=stream, kind=kind)
+    qr_code.save(out=stream, kind=kind, scale=scale)
     stream.seek(0)
     return stream
 
@@ -142,6 +179,17 @@ class FritzWLAN(AbstractLibraryBase):
     @ssid.setter
     def ssid(self, value):
         self._action('SetSSID', NewSSID=value)
+
+    @property
+    def beacontype(self):
+        """
+        Represents the beacontype for the network as string.
+        At time of writing (OS 7.29) possible values are:
+        `None, 11i, WPAand11i, 11iandWPA3` for the private WiFi-security
+        settings and `None, 11i, 11iandWPA3, OWETrans` for the guest
+        network.
+        """
+        return self.get_info()['NewBeaconType']
 
     @property
     def channel(self):